refactor(manager): injectuploadformをバラしてpostEmojiAddに含めた
